+++ include/slepcst.h
@@ -168,6 +168,32 @@ PETSC_DEPRECATED_FUNCTION(3, 15, 0, "STSetPreconditionerMat()", ) static inline 
 SLEPC_EXTERN PetscErrorCode STPrecondGetKSPHasMat(ST,PetscBool*);
 SLEPC_EXTERN PetscErrorCode STPrecondSetKSPHasMat(ST,PetscBool);
 
+/*E
+    STFilterType - Selects the method used to build the filter
+
+    Level: intermediate
+
+.seealso: STSetFilterType(), STGetFilterType()
+E*/
+typedef enum { ST_FILTER_FILTLAN   = 1,
+               ST_FILTER_CHEBYSHEV = 2 } STFilterType;
+SLEPC_EXTERN const char *STFilterTypes[];
+
+/*E
+    STFilterDamping - The damping type used to build the filter
+
+    Level: advanced
+
+.seealso: STSetFilterDamping(), STGetFilterDamping()
+E*/
+typedef enum { ST_FILTER_DAMPING_NONE,
+               ST_FILTER_DAMPING_JACKSON,
+               ST_FILTER_DAMPING_LANCZOS,
+               ST_FILTER_DAMPING_FEJER } STFilterDamping;
+SLEPC_EXTERN const char *STFilterDampings[];
+
+SLEPC_EXTERN PetscErrorCode STFilterSetType(ST,STFilterType);
+SLEPC_EXTERN PetscErrorCode STFilterGetType(ST,STFilterType*);
 SLEPC_EXTERN PetscErrorCode STFilterSetInterval(ST,PetscReal,PetscReal);
 SLEPC_EXTERN PetscErrorCode STFilterGetInterval(ST,PetscReal*,PetscReal*);
 SLEPC_EXTERN PetscErrorCode STFilterSetRange(ST,PetscReal,PetscReal);
@@ -175,3 +201,5 @@ SLEPC_EXTERN PetscErrorCode STFilterGetRange(ST,PetscReal*,PetscReal*);
 SLEPC_EXTERN PetscErrorCode STFilterSetDegree(ST,PetscInt);
 SLEPC_EXTERN PetscErrorCode STFilterGetDegree(ST,PetscInt*);
 SLEPC_EXTERN PetscErrorCode STFilterGetThreshold(ST,PetscReal*);
+SLEPC_EXTERN PetscErrorCode STFilterSetDamping(ST,STFilterDamping);
+SLEPC_EXTERN PetscErrorCode STFilterGetDamping(ST,STFilterDamping*);