+++ src/binding/slepc4py/src/slepc4py/SLEPc/ST.pyx
@@ -31,6 +31,30 @@ class STMatMode(object):
     INPLACE = ST_MATMODE_INPLACE
     SHELL   = ST_MATMODE_SHELL
 
+class STFilterType(object):
+    """
+    ST filter type
+
+    - `FILTLAN`:  An adapted implementation of the Filtered Lanczos Package.
+    - `CHEBYSEV`: A polynomial filter based on a truncated Chebyshev series.
+    """
+    FILTLAN   = ST_FILTER_FILTLAN
+    CHEBYSHEV = ST_FILTER_CHEBYSHEV
+
+class STFilterDamping(object):
+    """
+    ST filter damping
+
+    - `NONE`:    No damping
+    - `JACKSON`: Jackson damping
+    - `LANCZOS`: Lanczos damping
+    - `FEJER`:   Fejer damping
+    """
+    NONE    = ST_FILTER_DAMPING_NONE
+    JACKSON = ST_FILTER_DAMPING_JACKSON
+    LANCZOS = ST_FILTER_DAMPING_LANCZOS
+    FEJER   = ST_FILTER_DAMPING_FEJER
+
 # -----------------------------------------------------------------------------
 
 cdef class ST(Object):
@@ -39,8 +63,10 @@ cdef class ST(Object):
     ST
     """
 
-    Type         = STType
-    MatMode      = STMatMode
+    Type          = STType
+    MatMode       = STMatMode
+    FilterType    = STFilterType
+    FilterDamping = STFilterDamping
 
     def __cinit__(self):
         self.obj = <PetscObject*> &self.st
@@ -555,6 +581,31 @@ cdef class ST(Object):
         CHKERR( STCayleyGetAntishift(self.st, &sval) )
         return toScalar(sval)
 
+    def setFilterType(self, filter_type):
+        """
+        Sets the method to be used to build the polynomial filter.
+
+        Parameter
+        ---------
+        filter_type: `ST.FilterType` enumerate
+              The type of filter.
+        """
+        cdef SlepcSTFilterType val = filter_type
+        CHKERR( STFilterSetType(self.st, val) )
+
+    def getFilterType(self):
+        """
+        Gets the method to be used to build the polynomial filter.
+
+        Returns
+        -------
+        filter_type: `ST.FilterType` enumerate
+              The type of filter.
+        """
+        cdef SlepcSTFilterType val = ST_FILTER_FILTLAN
+        CHKERR( STFilterGetType(self.st, &val) )
+        return val
+
     def setFilterInterval(self, inta, intb):
         """
         Defines the interval containing the desired eigenvalues.
@@ -660,6 +711,31 @@ cdef class ST(Object):
         cdef PetscInt val = 0
         CHKERR( STFilterGetDegree(self.st, &val) )
         return toInt(val)
+
+    def setFilterDamping(self, damping):
+        """
+        Sets the type of damping to be used in the polynomial filter.
+
+        Parameter
+        ---------
+        damping: `ST.FilterDamping` enumerate
+              The type of damping.
+        """
+        cdef SlepcSTFilterDamping val = damping
+        CHKERR( STFilterSetDamping(self.st, val) )
+
+    def getFilterDamping(self):
+        """
+        Gets the type of damping used in the polynomial filter.
+
+        Returns
+        -------
+        damping: `ST.FilterDamping` enumerate
+              The type of damping.
+        """
+        cdef SlepcSTFilterDamping val = ST_FILTER_DAMPING_NONE
+        CHKERR( STFilterGetDamping(self.st, &val) )
+        return val
 
     #